+++ neat/model_gc_bias/__init__.py
@@ -1,1 +1,2 @@
 from .runner import compute_gc_bias_runner
+from .utils import compute_genome_wide_gc_mean_weight